# backend/openapi/v1.json
{
  "components": {
    "schemas": {
      "Accepted": {
        "description": "What a call answers when the caller asked not to wait: the task it started.",
        "properties": {
          "job_id": {
            "description": "Follow it with `GET /jobs/{job_id}`, whose `result` holds the report.",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "The envelope every refusal and failure answers.",
        "properties": {
          "answerable": {
            "description": "Beside `confirm`: whether this caller may send the name back.",
            "type": [
              "boolean",
              "null"
            ]
          },
          "confirm": {
            "description": "The guardrail that asks, for `confirmation_required`. Sending it back\nin `confirm` accepts that question and no other.",
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "description": "A stable code: `bad_request`, `unauthorized`, `forbidden`, `not_found`,\n`method_not_allowed`, `conflict`, `confirmation_required`,\n`payload_too_large`, `unsupported_media_type`, `too_many_requests`,\n`busy`, `reauthentication_required`, `passphrase_required`,\n`external_api_error` or `internal_error`. An open list.",
            "type": "string"
          },
          "includes": {
            "description": "Beside `confirm`: the other guardrails the question states, sent back\nin `confirm` with it to accept the question whole.",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "message": {
            "description": "A sentence for a person, in the interface language. Never part of the\ncontract.",
            "type": "string"
          },
          "reason": {
            "description": "Why a `conflict` refuses, an open list. A job still running, with a\n`Retry-After` header: `apply_running`, `simulation_running`,\n`previews_full`, `sync_running`, `enrichment_running`, `backup_running`,\n`backup_just_taken`, `restore_staging`, `purge_running`. A state that\nstays until somebody changes it: `name_taken`, `in_use`,\n`already_listed`, `not_declared`, `category_mapped`,\n`proposals_replaced`, `dry_run`, `not_cancellable`, `key_pinned`,\n`only_credential`, `secret_unreadable`.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "error",
          "message"
        ],
        "type": "object"
      },
...
      "InstanceResponse": {
        "description": "An Arr Routarr reads.",
        "properties": {
          "api_key_encrypted": {
            "description": "Whether the stored API key is encrypted at rest.",
            "type": "boolean"
          },
          "api_key_masked": {
            "description": "The Arr's key, masked. Empty for an application key.",
            "type": "string"
          },
          "base_url": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "instance_type": {
            "type": "string"
          },
          "last_sync_at": {
            "description": "When a sync last *succeeded*. Null while none ever has.",
            "type": [
              "string",
              "null"
            ]
          },
          "last_sync_attempt_at": {
            "description": "When one was last attempted, successful or not: the figure that says\nwhether the scheduler is running at all.",
            "type": [
              "string",
              "null"
            ]
          },
          "last_sync_status": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "sync_interval_minutes": {
            "format": "int64",
            "type": "integer"
          },
          "updated_at": {
            "type": "string"
          },
          "webhook_path": {
            "description": "The same without the token, which goes in `X-Routarr-Token`: Radarr\n5.16 and Sonarr 4.0.11 send custom headers. `null` for an application\nkey.",
            "type": [
              "string",
              "null"
            ]
          },
          "webhook_token": {
            "description": "The token for `X-Routarr-Token`. `null` for an application key.",
            "type": [
              "string",
              "null"
            ]
          },
          "webhook_url": {
            "description": "Relative URL to register in Radarr/Sonarr's webhook connection, the\ntoken in it, for an Arr that sends no custom header. `null` for an\napplication key: its token lets anyone post events as the Arr.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "instance_type",
          "base_url",
          "api_key_masked",
          "enabled",
          "sync_interval_minutes",
          "created_at",
          "updated_at",
          "api_key_encrypted"
        ],
        "type": "object"
      },
...
      "SyncReport": {
        "description": "Result of syncing one instance.",
        "properties": {
          "error": {
            "description": "Set when the sync failed, so the caller of sync-all still gets one entry\nper instance instead of the failure silently vanishing from the report.",
            "type": [
              "string",
              "null"
            ]
          },
          "instance_id": {
            "type": "string"
          },
          "instance_name": {
            "type": "string"
          },
          "media": {
            "minimum": 0,
            "type": "integer"
          },
          "removed": {
            "description": "Media rows that disappeared upstream and were removed locally.",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "root_folders": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "instance_id",
          "instance_name",
          "root_folders",
          "media",
          "removed"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "in": "header",
        "name": "X-Api-Key",
        "type": "apiKey"
      },
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "license": {
      "identifier": "GPL-3.0-only",
      "name": "GPL-3.0-only"
    },
    "title": "Routarr API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/instances": {
      "get": {
        "description": "Their address, kind and last sync. Never the Arr's key, and for an\napplication key neither the webhook address, whose token lets anyone post\nevents as the Arr: `webhook_url` is `null` and `api_key_masked` empty.",
        "operationId": "list_instances",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InstanceResponse"
                  },
                  "type": "array"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A refusal or a failure, in the envelope every error shares."
          }
        },
        "summary": "The Arrs Routarr reads",
        "tags": [
          "instances"
        ],
        "x-routarr-scope": "read"
      }
    },
    "/instances/sync": {
      "post": {
        "description": "Answers once every instance is read, one report each. An instance that\nfails is reported with its `error` and stops none of the others. With\n`respond-async` it answers once its task has started, and that task holds\nthe reports when it ends. Each instance is read under a task of its own too.",
        "operationId": "sync_all",
        "parameters": [
          {
            "description": "`respond-async` answers 202 once the task has started, instead of its reports.",
            "in": "header",
            "name": "Prefer",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SyncReport"
                  },
                  "type": "array"
                }
              }
            }
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            },
            "description": "The task has started, as `Prefer: respond-async` asked.",
            "headers": {
              "Location": {
                "description": "The task, under `/jobs`.",
                "schema": {
                  "type": "string"
                }
              },
              "Preference-Applied": {
                "description": "`respond-async`.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Seconds to wait before the first look at the task.",
                "schema": {
                  "format": "int32",
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A refusal or a failure, in the envelope every error shares."
          }
        },
        "summary": "Read every enabled Arr again",
        "tags": [
          "instances"
        ],
        "x-routarr-scope": "operate"
      }
    },
    "/instances/{id}": {
      "get": {
        "operationId": "get_instance",
        "parameters": [
          {
            "description": "The instance's id in Routarr.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceResponse"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A refusal or a failure, in the envelope every error shares."
          }
        },
        "summary": "One Arr",
        "tags": [
          "instances"
        ],
        "x-routarr-scope": "read"
      }
    },
    "/instances/{id}/sync": {
      "post": {
        "description": "A sync of an instance already running answers 409.",
        "operationId": "sync_instance",
        "parameters": [
          {
            "description": "The instance's id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "`respond-async` answers 202 once the task has started, instead of its report.",
            "in": "header",
            "name": "Prefer",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncReport"
                }
              }
            }
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Accepted"
                }
              }
            },
            "description": "The task has started, as `Prefer: respond-async` asked.",
            "headers": {
              "Location": {
                "description": "The task, under `/jobs`.",
                "schema": {
                  "type": "string"
                }
              },
              "Preference-Applied": {
                "description": "`respond-async`.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Seconds to wait before the first look at the task.",
                "schema": {
                  "format": "int32",
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A refusal or a failure, in the envelope every error shares."
          }
        },
        "summary": "Read one Arr again",
        "tags": [
          "instances"
        ],
        "x-routarr-scope": "operate"
      }
    },
...
  },
  "security": [
    {
      "api_key": []
    },
    {
      "bearer": []
    }
  ],
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}